packages/sanjaya/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sanjaya"
version = "0.1.0"
description = "Django Ninja app for dynamic reporting with pluggable data providers"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = "MIT"
dependencies = [
    "django>=4.2,<6",
    "django-ninja>=1.0,<2",
    "sanjaya-core>=0.1.0",
    "openpyxl>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-django>=4.5",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sanjaya"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
DJANGO_SETTINGS_MODULE = "tests.settings"
